Update index.mdx (#114)
Have made some changes here. I think the steps are currently a bit confusing. We mention the exporter and in other places call them the manufacturer so I have changed it to match that. I've also got rid of the part about land transport and loading etc as this is not implemented. Still need to repoint the reefer.png image and add in one for the scenario overview.